removed 'plot_' prefix and added necessary files
