+++ examples/run_ilastik.sh
@@ -1,0 +1,1 @@
+THIS IS A FAKE SHELL SCRIPT INTENDED TO TRICK :py:class:`.IlastikPretrainedProbability`
